requirements.txt
fastapi
uvicorn[standard]
pydantic==2.12.4
numpy
pandas
scikit-learn
lightgbm
mlflow
joblib
pyyaml
